+++ src/main/java/moze_intel/projecte/gameObjs/tiles/CondenserTile.java
@@ -25,11 +25,11 @@ public class CondenserTile extends TileEmcDirection implements IInventory, ISide
 	protected boolean loadChecks;
 	protected boolean isAcceptingEmc;
 	private int ticksSinceSync;
-	public int displayEmc;
+	public double displayEmc;
 	public float lidAngle;
 	public float prevLidAngle;
 	public int numPlayersUsing;
-	public int requiredEmc;
+	public long requiredEmc;
 
 	public CondenserTile()
 	{
@@ -54,7 +54,7 @@ public class CondenserTile extends TileEmcDirection implements IInventory, ISide
 			loadChecks = true;
 		}
 
-		displayEmc = (int) this.getStoredEmc();
+		displayEmc = this.getStoredEmc();
 
 		if (lock != null && requiredEmc != 0)
 		{
@@ -86,8 +86,7 @@ public class CondenserTile extends TileEmcDirection implements IInventory, ISide
 
 			if (requiredEmc != lockEmc)
 			{
-				//TODO: long
-				requiredEmc = (int) lockEmc;
+				requiredEmc = lockEmc;
 				this.isAcceptingEmc = true;
 			}
 		}
@@ -223,7 +222,7 @@ public class CondenserTile extends TileEmcDirection implements IInventory, ISide
 			return Constants.MAX_CONDENSER_PROGRESS;
 		}
 		
-		return (displayEmc * Constants.MAX_CONDENSER_PROGRESS) / requiredEmc;
+		return (int) ((displayEmc * Constants.MAX_CONDENSER_PROGRESS) / requiredEmc);
 	}
 
 	@Override
